requirements-dev.txt
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
ipython==9.9.0
pillow==12.1.0
pip==25.3
psycopg2-binary==2.9.11
